requirements.txt
asgiref==3.8.1
Django==4.2.16
django-ckeditor-5==0.2.15
django-colorfield==0.11.0
django-cors-headers==4.7.0
django-js-asset==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
pillow==11.0.0
python-decouple==3.8
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.12.2
uritemplate==4.1.1
